Cleanup nest auth registration, moving into config_flow
Move the nest auth handlers and registry into config_flow.py, a pre-factoring
step for a future PR. In the future, the auth flows will be dynamically
created in the middle of the config flow following https://github.com/home-assistant/architecture/discussions/692
and this avoids a circular dependency and matches the structure
for the legacy works with nest flow anyway.
